% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catData.R
\name{catData}
\alias{catData}
\title{Categorize continuous data}
\usage{
catData(data, k = 4, sym = "sym", standardize = FALSE)
}
\arguments{
\item{data}{A numerical data.frame, matrix or vector to be categorized.}

\item{k}{The number of categories to divide the data into. An integer
between 2 and 7.}

\item{sym}{One of \code{c("sym", "moderate", "extreme", "moderate-alt",
"extreme-alt")}. By default (\code{sym}), the data is divided into equally
distanced cut-off points between -2.5 and 2.5. In \code{moderate}
assymmetry, cut-offs are chosen so that the peak of a normal distribution
falls on the left-hand side. \code{extreme} assymmetry results in a
distribution where the first category has the highest number of cases.
'alt' versions of the distributions invert the cut-off points.}

\item{standardize}{Boolean. Should parameters be standardized before
transformation?}
}
\description{
Transforms continuous data into categorical data, based on Z-score cut-off
points.
}
