% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_ML.R
\name{cfa_ML}
\alias{cfa_ML}
\title{Run the ML models}
\usage{
cfa_ML(run_id, id, models, cat_data, ...)
}
\arguments{
\item{run_id}{unique identifier for each run}

\item{id}{unique identifier for each specified model}

\item{models}{string indicating model object "model1" or "model2"}

\item{cat_data}{categorical data set}
}
\description{
Runs models using Maximum Likelihood (ML) parameter estimation.
}
