% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_ML.R
\name{cfa_ULS}
\alias{cfa_ULS}
\title{Run the ULS models}
\usage{
cfa_ULS(run_id, id, models, cat_data)
}
\arguments{
\item{run_id}{unique identifier for each run}

\item{id}{unique identifier for each specified model}

\item{models}{model file}

\item{cat_data}{categorical data set}
}
\description{
Runs models using Unweighted Least Squares (ULS) parameter estimation.
}
