% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatData.R
\name{getCatData}
\alias{getCatData}
\title{Generate categorical data}
\usage{
getCatData(models, N, cat, sym, dist, seed = 42, ...)
}
\arguments{
\item{models}{}

\item{N}{}

\item{cat}{}

\item{sym}{}

\item{dist}{}

\item{seed}{}

\item{...}{}
}
\value{

}
\description{
Generate categorical data
}
