% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Generate data from a SEM model}
\usage{
getData(
  N = 100,
  loadings = NULL,
  population = NULL,
  latent.cor = NULL,
  dist = NULL,
  model = NULL
)
}
\arguments{
\item{N}{The number of cases that should be generated.}

\item{loadings}{Matrix with number of rows equal to the number of
observed variables, and number of collumns equal to the number of latent
factors. Values specified are fixed factor loadings in the model. \code{NA}
values indicate freed parameters, and will be estimated.}

\item{population}{Matrix with identical dimensions to the loadings matrix.
Indicates population values for data generation.}

\item{latent.cor}{Vector of latent factor correlations in the population
(estimated by default).}

\item{dist}{Population distribution. Use the \code{bindDist}
function to specify a distribution. By default (\code{NULL} data is drawn
from a normal distribution.)}

\item{model}{Supply any custom model of class \code{SimSem}. When
used, bypasses the loadings, population and latent.cor values.}
}
\description{
A wrapper function around \code{generate} from the package \code{SimSem}.
Creates data from a CFA model with normal or non-normal data.
}
\examples{
# generate a matrix from a 2 factor CFA model with 10 observed variables.
loadings <- matrix(0, 10, 2)
loadings[1:5, 1] <- NA
loadings[6:10, 2] <- NA
pop <- matrix(0, 10, 2)
pop[1:5, 1] <- c(.3, .4, .5, .6, .7)
pop[6:10, 2] <- c(.3, .4, .5, .6, .7)

dist <- simsem::bindDist(skewness = 2, kurtosis = 7)
data <- getData(loadings = loadings, population = pop, latent.cor = 0.3, dist = dist)
}
