% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimates.R
\name{getEstimates}
\alias{getEstimates}
\title{Extract relevant parameters from a fitted Lavaan model.}
\usage{
getEstimates(modfit, f = 2)
}
\arguments{
\item{modfit}{a fitted model object}
}
\description{
Internal function to extract parameters of interest from a fitted model.
}
