% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScenarios.R
\name{getScenarios}
\alias{getScenarios}
\title{Simulate model outcomes under various conditions}
\usage{
getScenarios(
  its = 100,
  N = 100,
  cat = 5,
  sym = "sym",
  models,
  dist = NULL,
  conditions = NULL
)
}
\arguments{
\item{its}{Number of iterations to simulate for.}

\item{N}{Vector of integers representing the number of cases for
which data is simulated.}

\item{cat}{A vector of integers between 2 and 7. Number of
categories to divide the data into.}

\item{sym}{Character vector containing one or more of
\code{c("sym", "moderate", "extreme", "moderate-alt", "extreme-alt")}.
Symmetry of the categorical divisions.}

\item{models}{Vector containing \code{SimSem} model files. Can be
generated with the \code{getModel} wrapper function.}

\item{dist}{Vector of \code{SimDataDist} objects. Specifies the
distributions used to draw the data from. Note that the categorization
always assumes a normal distribution.}

\item{conditions}{Tibble file containing all conditions for which a user
wants to simulate. Bypasses the parameters \code{N}, \code{cat},
\code{sym}, \code{models} and \code{dist}.}
}
\description{
This function generates a \code{data.frame} object containing all scenarios
to simulate.
}
