% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhemcon.R
\docType{data}
\name{rhemcon}
\alias{rhemcon}
\title{Rhemtulla, Brosseau-Liard & Savalei (2012) simulation conditions}
\format{
A tibble file containing the 480 conditions under which Rhemtulla
  Brosseau-Liard & Savalei conducted their simulation study.
}
\description{
Rhemtulla, Brosseau-Liard & Savalei (2012) simulation conditions
}
